+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     transaction messages for CreateDenom, ChangeAdmin, and UpdateModuleParams 
   * [#1620](https://github.com/NibiruChain/nibiru/pull/1620) - Token factory
     transaction messages for Mint and Burn
+* [#1573](https://github.com/NibiruChain/nibiru/pull/1573) - feat(perp): Close markets and compute settlement price
 
 ### State Machine Breaking